requirements.txt
redis
pycrypto
Flask
Flask-RESTful
Flask-HTTPAuth
requests
SimpleConfigParser
